Migrate protocol, units, and results for the hybridtop protocol
