+++ openfe/protocols/openmm_rfe/__init__.py
@@ -2,11 +2,7 @@
 # For details, see https://github.com/OpenFreeEnergy/openfe
 
 from . import _rfe_utils
-from .equil_rfe_methods import (
-    RelativeHybridTopologyProtocol,
-    RelativeHybridTopologyProtocolResult,
-    RelativeHybridTopologyProtocolUnit,
-)
-from .equil_rfe_settings import (
-    RelativeHybridTopologyProtocolSettings,
-)
+from .hybridtop_protocols import RelativeHybridTopologyProtocol
+from .hybridtop_unit_results import RelativeHybridTopologyProtocolResult
+from .hybridtop_units import RelativeHybridTopologyProtocolUnit
+from .equil_rfe_settings import RelativeHybridTopologyProtocolSettings
